feat(Theme): 관리자 테마 삭제 API 예외 처리 추가

+++ src/main/java/roomescape/domain/theme/service/ThemeService.java
@@ -58,6 +58,10 @@ public class ThemeService {
     }
 
     public void deleteThemeById(Long id) {
+        if (!themeRepository.existsThemeByIdAndDeletedAtIsNull(id)) {
+            throw new GeneralException(ThemeErrorType.THEME_NOT_FOUND);
+        }
+
         themeRepository.deleteThemeById(id);
     }
 }
